+++ src/fetch/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 Note for all Flow users: all projects in [`adeira/universe`](https://github.com/adeira/universe) now use implicit exact Flow types (`{}` for strict objects and `{ ... }` for open objects, syntax `{||}` is deprecated). We do not expect any issues as long as you are using `exact_by_default=true` Flow option.
 
+- Do not retry for POST requests with [`FormData`](https://developer.mozilla.org/en-US/docs/Web/API/FormData) body (for example GraphQL uploads).
+
 # 2.0.2
 
 Dependency `cross-fetch` update: